revno: 1965 message:   Fixed bug related to mysql double type
committer: Lars Helge Overland <larshelge@gmail.com>
branch nick: dhis2
timestamp: Fri 2010-10-29 16:22:06 +0200

+++ dhis-2/dhis-support/dhis-support-jdbc/src/main/java/org/hisp/dhis/jdbc/statementbuilder/MySQLStatementBuilder.java
@@ -41,7 +41,7 @@ public class MySQLStatementBuilder
 {
     public String getDoubleColumnType()
     {
-        return "DOUBLE";
+        return "DECIMAL";
     }
 
     public String encode( String value )
